refactor: extract ConfigHistorySection and preambles to fix file size CI check
DomainSettingsPanel.tsx was 466 lines, over the 400-line limit. Extracted
the config history block into ConfigHistorySection.tsx and the inline
PREAMBLES_QUERY + types into api/operations/preambles.ts, bringing the
panel down to 382 lines.
